+++ lib/src/universal_platform.dart
@@ -1,0 +1,3 @@
+export 'universal_platform_none.dart'
+    if (dart.library.io) 'universal_platform_io.dart'
+    if (dart.library.html) 'universal_platform_html.dart';