+++ lib/src/universal_platform_type.dart
@@ -1,0 +1,11 @@
+enum UniversalPlatformType {
+  Web,
+  Windows,
+  Linux,
+  MacOS,
+  Android,
+  Fuchsia,
+  IOS,
+  Other,
+  None,
+}