+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Changelog
+
+## 1.1.0
+
+- Refactor to call Platform.is* in a more direct way
+- Make UniversalPlatform.isWeb constant
+- Add UniversalPlatformType.Other and .None
+
 ## 1.0.0-nullsafety
 - Migrate to nnbd